FILE: renames create-places-tree to create-places-root
